refactor(settings): split into module hierarchy for single-responsibility
